requirements.txt
anyio==3.6.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
exceptiongroup==1.1.0
fastapi==0.88.0
h11==0.14.0
idna==3.4
iniconfig==2.0.0
packaging==23.0
pluggy==1.0.0
psycopg2-binary==2.9.5
pydantic==1.10.4
pytest==7.2.0
python-dotenv==0.21.0
requests==2.28.1
sniffio==1.3.0
SQLAlchemy==1.4.46
starlette==0.22.0
tomli==2.0.1
typing_extensions==4.4.0
urllib3==1.26.13
uvicorn==0.20.0
